requirements.txt
Flask
Werkzeug
biopython
numpy
pandas